Add `FixProjectileFF` feature
This feature addresses the bug that allows teammates to explode some of
the player's projectiles by damaging them even when friendly fire is
turned off, therefore killing the player (whether by accident or not).

Problem is solved by "disarming" projectiles vulnerable to this friendly
fire and replacing them with our own class of projectile that is spawned
only on a server and does additional safety checks to ensure it will
only explode when it is expected from it.