Remove TODO about certain code duplication
This is about code that looks through player's inventory for items of a
particular classes and similar code being used in several other places.

While `AcediaCore` now comes with an API method that does precisely
that, the algorithm itself is idiomatic enough to not be bothered with
it being rewritten in sevral different places and trying to replace
every such piece of code simply for getting rid of code duplication can
lead to introduction of new bugs with little to no actual benefit to
maintability.